pyproject.toml
[project]
name = "qiskit-runtime-server"
version = "0.1.0"
description = "Self-hosted IBM Qiskit Runtime compatible REST API server"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "gyu-don", email = "takumi.kt@gmail.com" }
]
keywords = [
    "qiskit",
    "quantum",
    "runtime",
    "server",
    "api",
    "simulation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Physics",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.123.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "qiskit-aer>=0.17.2",
    "qiskit-ibm-runtime>=0.43.1",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
qiskit-runtime-server = "qiskit_runtime_server.__main__:main"

[project.urls]
Homepage = "https://github.com/gyu-don/qiskit-runtime-server"
Documentation = "https://github.com/gyu-don/qiskit-runtime-server#readme"
Repository = "https://github.com/gyu-don/qiskit-runtime-server"
Issues = "https://github.com/gyu-don/qiskit-runtime-server/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qiskit_runtime_server"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["qiskit_runtime_server"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "qiskit.*",
    "qiskit_ibm_runtime.*",
    "qiskit_aer.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
#fail_under = 80
show_missing = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "types-requests>=2.32.4.20250913",
]
